batches: Address race condition for search result paths
Indexed search is constantly playing catch up with commits being pushed to repos. Now, when I add a new commit, it takes a couple seconds to minutes (depending on the repo) until the latest commit is actually indexed and hence searched when running a search. We knew there can be discrepancies between search results when the preview ran and when the execution starts, so we do persist the commit hash along with the search result paths, but we do not take the commit from the search result, we simply take latest commit on HEAD. So there’s this discrepancy between commits while zoekt indexes, where running the search and fetching latest head for a repo returns results for different states.

That can likely be fixed - played around with it a little in this PR.
But it also raised two more questions - what if there are multiple searchable branches configured for a repo, and what if a user does `r:repo@rev`. Those cases likely also won’t work with search result paths today.
More investigation to be done!